+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 <!-- CHANGELOG NEWEST -->
 - FIX firestore replication not working with schema validation [#5572](https://github.com/pubkey/rxdb/issues/5572)
+- FIX Replication observation mode ignored when push handler is waiting [#5571](https://github.com/pubkey/rxdb/issues/5571)
+
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->